Revert r827377 (the commit that reverted the kleo merge). Now everything is back to the way it was before I started this mess today.
The problem I couldn't overcome was that svn diff / patch don't deal with
deleted,added and moved files properly.

Maybe I'll start another attempt using svn merge (not svnmerge), but not
today.

svn path=/trunk/KDE/kdepim/kleopatra/uiserver/; revision=827403

+++ libkleo/ui/keyapprovaldialog.cpp
@@ -192,7 +192,7 @@ std::vector<GpgME::Key> Kleo::KeyApprovalDialog::senderKeys() const {
 }
 
 std::vector<Kleo::KeyApprovalDialog::Item> Kleo::KeyApprovalDialog::items() const {
-  assert( d->requesters.size() == d->addresses.size() );
+  assert( d->requesters.size() == static_cast<unsigned int>( d->addresses.size() ) );
   assert( d->requesters.size() == d->preferences.size() );
 
   std::vector<Item> result;
